EQREL: Merge 'extend' with insertion
This is intended to help with performance. Previously 'extend' was
called, which adds (up to) all the tuples from the old relation to the
new relation, right before inserting all the tuples from the new
relation back into the old relation. But simply switching the order of
insertion and extend still isn't optimal - Then the old relation would
contain a copy of the new one when 'extend' is called. So instead,
combined the two operations.

Furthermore, insertion previously used the "relation" interface to the
EQREL, which may be less performant than using the "eqrel" interface as
in this commit.

See PR #2071 for discussion.